Fix worker import script
