+++ scripts/import_workers.ts
@@ -1,6 +1,4 @@
-#!/usr/bin/env bun
 import { invoke } from '@tauri-apps/api/tauri';
-import { readFileSync } from 'fs';
 
 export async function importWorkers(content: string, token: string = '') {
   const workers = content
@@ -11,16 +9,8 @@ export async function importWorkers(content: string, token: string = '') {
   return workers.length;
 }
 
-async function main() {
-  const file = process.argv[2];
-  const token = process.argv[3] ?? '';
-  if (!file) {
-    console.error('Usage: import_workers.ts <file> [token]');
-    process.exit(1);
-  }
-  const content = readFileSync(file, 'utf-8');
-  const count = await importWorkers(content, token);
-  console.log(`Imported ${count} workers`);
+export async function importWorkersFromFile(path: string, token = '') {
+  const { readFileSync } = await import('fs');
+  const content = readFileSync(path, 'utf-8');
+  return importWorkers(content, token);
 }
-
-main();
